refactor: move files related to onboarding screens to their correct package
