+++ domain/src/main/java/com/amsterdam/domain/useCase/preferences/GetOnboardingStatusUseCase.kt
@@ -9,10 +9,3 @@ class GetOnboardingStatusUseCase(
         return preferencesRepository.isOnboardingCompleted()
     }
 }
-class SetOnboardingCompletedUseCase(
-    private val preferencesRepository: AppPreferencesRepository
-) {
-    suspend operator fun invoke(isCompleted: Boolean) {
-        preferencesRepository.setOnboardingCompleted(isCompleted)
-    }
-}
