+++ ui/src/main/java/com/amsterdam/ui/screens/onBoarding/OnBoardingPage.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.amsterdam.designsystem.components.PageIndicator
+import com.amsterdam.ui.screens.onBoarding.component.PageIndicator
 import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AppTheme
