+++ ui/src/main/java/com/amsterdam/ui/screens/onBoarding/component/NavigationButton.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.designsystem.components.buttons
+package com.amsterdam.ui.screens.onBoarding.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -8,8 +8,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ripple
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -19,9 +17,11 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.amsterdam.designsystem.R
+import com.amsterdam.designsystem.components.Icon
 import com.amsterdam.designsystem.theme.AflamiTheme
 import com.amsterdam.designsystem.theme.AppTheme
 import com.amsterdam.designsystem.utils.ThemeAndLocalePreviews
+import com.amsterdam.designsystem.utils.ripple
 
 @Composable
 fun NavigationButton(
