+++ ui/src/main/java/com/amsterdam/ui/screens/onBoarding/component/SkipButton.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.designsystem.components.buttons
+package com.amsterdam.ui.screens.onBoarding.component
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.tween
@@ -9,7 +9,6 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -20,6 +19,7 @@ import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import com.amsterdam.designsystem.R
+import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AppTheme
 
 @Composable
